use plaintext for did and append tags with @ symbol
reason for using plaintext is that it forces me to keep things simple
which is probably better for the long run. using markdown will make me
want to add tables and images and all sorts of stuff but it's really not
necessary to format the notes that much if they're all just going to sit
in one big file.

as for prefacing the tags with an @symbol, that's just to make them
easier to search for when i'm grepping the 'did' file