make firefox.sh more robust
it didn't work when I tested it on MacOS, and I can't be bothered to
make it work across platforms, so to save some time, I've added a bit of
code to the top of the script which forces it to exit if it detects that
it's not running on Linux.

there are also some extra checks that happen to ensure that you've got
all of the required binaries installed before it tries to do anything
(so it doesn't fail halfway and leave your files in a weird state).