Fix broken install script
It was broken because I moved it to another folder but I didn't move the
requirements files with it, so it would fail to find them when you ran
the script.

Didn't notice until now because the 'installation' part of the script
gets skipped if you don't actually need to install anything.