migrate to dotbot
now, my dotfiles are installed automatically and managed by the
install.conf.yaml instead of that hacky shell script.

dotbot will create symlinks, so the config files can exist in the repo
and at various places in the OS at the same time.

dotbot will also manage the dead symlinks (if any) that appear in the
home dir (might need to configure this a bit, not sure)

all in all, the config file management is a lot cleaner now.