update workstation nb fzf (use sk and ripgrep)
Now, instead of fzf, we're using skim and ripgrep to do the searching
behind the scenes. With this change, we can now search the contents of
the notes, and not the just title, which makes this a proper full-text
search now.

The plugin is still called via 'nb fzf' because I find that easier to
type.

We're using the preview.sh script from below to generate previews
https://github.com/junegunn/fzf.vim/blob/master/bin/preview.sh

closes gh-30