update dotfile installation yaml file
found out that it wasn't cleaning up the broken symlinks underneath the
HOME dir, so for now I've added all of the directories to the 'clean'
section of the config.

Maybe I could just make it so that it recursively searches the HOME dir?
Haven't tested the performance of that though, so don't know how well
that would work.