add ansible tasks for installing 'bin'
bin is a manager for those standalone binaries that most projects
include in their GitHub releases page.

since I can't use the package manager to install stuff on Silverblue I
am using bin to get the binaries into my .local/bin folder.

bin will keep track of the versions as well, which will make it easier
to do updates. The only problem is that it's difficult to manage bin
with Ansible because it doesn't work well in a headless fashion. I am
following a couple of issues that aim to resolve this but there haven't
been any updates on those for a while, so for now Ansible will only
handle installing bin - installing the apps themselves will have to be
done in a shell script somewhere (because it will need user input).

closes gh-38
ref gh-16