install fzf if not already present
the snippet that installs fzf is part of the bash init scripts because
it was easier to put it there than to put it into the playbook.

it's only two lines of code and it's guarded by a check to ensure that
it only runs when fzf is NOT present (& when ghq _is_ present) and I
have tested this, so happy to call it a day here; documentation has also
been updated.

closes gh-40