update workstation manual setup docs
I've installed a new PWA via Brave; this commit updates that part of the
documentation to reflect this (still not sure how to automate the
installation of PWAs so this is as good as it'll be for now).

Fleshed out the TODO note on the nb setup section, because that setup
should be automated and I think the easiest way would be to follow what
I've done for pipx and bin.

And to finish it off, this commit removes the docs for pipx apps because
that's now in the group_vars/local.yml part of the playbook folder and
it's managed by the bash init scripts (which are installed by the
playbook).