create 'apps-maintain' script
The new script is generated by the Ansible playbook and placed in the
local bin folder.

The bash init scripts hook into this script and run it once per boot.
Apps are considered stale based on whatever is defined in the Ansible
vars. If a 'number of days' is not configured for an app, it won't be
included in the script when Ansible generates it.

The script won't actually update anything, but it should put something
in STDOUT that lets the user know that which apps are stale and how old
they are.

closes gh-45