lots of changes
The playbook Makefile has been re-written and should be run from within
a toolbox container.

Install 'psutil' on host system because we can't use the venv to run
Ansible now that we are connecting 'remotely' from the container.

Make the documentation more succint and update it to mention exactly how
to use the new bootstrap0.sh script.

refs gh-63