fix brave policy issue
basically, my workaround is to use Ansible to insert the snippet back
into the brave startup script which was removed by the flatpak
maintainers.

this way, Brave will pick up the policy as before - the downside being
that I have to maintain this extra snippet.

might be worth contributing it back upstream, but it doesn't look like
they want it (given that they removed it from their codebase to fix some
bugs).

closes #106